Make settings configurable

+++ src/main/scala/eu/joaocosta/coffee/model/Settings.scala
@@ -5,6 +5,6 @@ import scala.concurrent.duration.*
 final case class Settings(
   caffeineHalfLife: FiniteDuration = 5.hours,
   caffeineIngestion: FiniteDuration = 30.minutes,
-  minCaffeine: Double = 100,
-  maxCaffeine: Double = 400
+  minCaffeine: Int = 100,
+  maxCaffeine: Int = 400
 )
